fixes estimate_memory, shifts BaseKeyedVectors to keyedvectors.py
