+++ gensim/models/keyedvectors.py
@@ -73,9 +73,8 @@ except ImportError:
 
 from numpy import dot, zeros, float32 as REAL,\
     double, array, vstack, sqrt, newaxis,\
-    ndarray, sum as np_sum, prod
+    ndarray, sum as np_sum, prod, argmax
 import numpy as np
-from gensim.models.base_any2vec import BaseKeyedVectors
 from gensim import utils, matutils  # utility fnc for pickling, common scipy operations etc
 from gensim.corpora.dictionary import Dictionary
 from six import string_types
@@ -102,6 +101,90 @@ class Vocab(object):
     def __str__(self):
         vals = ['%s:%r' % (key, self.__dict__[key]) for key in sorted(self.__dict__) if not key.startswith('_')]
         return "%s(%s)" % (self.__class__.__name__, ', '.join(vals))
+
+
+class BaseKeyedVectors(utils.SaveLoad):
+
+    def __init__(self):
+        self.vectors = []
+        self.vocab = {}
+        self.vector_size = None
+        self.index2entity = []
+
+    def save(self, fname_or_handle, **kwargs):
+        super(BaseKeyedVectors, self).save(fname_or_handle, **kwargs)
+
+    @classmethod
+    def load(cls, fname_or_handle, **kwargs):
+        return super(BaseKeyedVectors, cls).load(fname_or_handle, **kwargs)
+
+    def similarity(self, entity1, entity2):
+        """Compute cosine similarity between entities, specified by string tag.
+        """
+        raise NotImplementedError
+
+    def most_similar(self, **kwargs):
+        """Find the top-N most similar entities.
+        Possibly have `positive` and `negative` list of entities in `**kwargs`.
+        """
+        return NotImplementedError
+
+    def distance(self, entity1, entity2):
+        """Compute distance between vectors of two input entities, specified by string tag.
+        """
+        raise NotImplementedError
+
+    def distances(self, entity1, other_entities=()):
+        """Compute distances from given entity (string tag) to all entities in `other_entity`.
+        If `other_entities` is empty, return distance between `entity1` and all entities in vocab.
+        """
+        raise NotImplementedError
+
+    def get_vector(self, entity):
+        """Accept a single entity as input, specified by string tag.
+        Returns the entity's representations in vector space, as a 1D numpy array.
+        """
+        if entity in self.vocab:
+            result = self.vectors[self.vocab[entity].index]
+            result.setflags(write=False)
+            return result
+        else:
+            raise KeyError("'%s' not in vocabulary" % entity)
+
+    def __getitem__(self, entities):
+        """
+        Accept a single entity (string tag) or list of entities as input.
+
+        If a single string or int, return designated tag's vector
+        representation, as a 1D numpy array.
+
+        If a list, return designated tags' vector representations as a
+        2D numpy array: #tags x #vector_size.
+        """
+        if isinstance(entities, string_types):
+            # allow calls like trained_model['office'], as a shorthand for trained_model[['office']]
+            return self.get_vector(entities)
+
+        return vstack([self.get_vector(entity) for entity in entities])
+
+    def __contains__(self, entity):
+        return entity in self.vocab
+
+    def most_similar_to_given(self, entity1, entities_list):
+        """Return the entity from entities_list most similar to entity1."""
+        return entities_list[argmax([self.similarity(entity1, entity) for entity in entities_list])]
+
+    def closer_than(self, entity1, entity2):
+        """Returns all entities that are closer to `entity1` than `entity2` is to `entity1`."""
+        all_distances = self.distances(entity1)
+        e1_index = self.vocab[entity1].index
+        e2_index = self.vocab[entity2].index
+        closer_node_indices = np.where(all_distances < all_distances[e2_index])[0]
+        return [self.index2entity[index] for index in closer_node_indices if index != e1_index]
+
+    def rank(self, entity1, entity2):
+        """Rank of the distance of `entity2` from `entity1`, in relation to distances of all entities from `entity1`."""
+        return len(self.closer_than(entity1, entity2)) + 1
 
 
 class WordEmbeddingsKeyedVectors(BaseKeyedVectors):
